[Tizen] Refactor the TabbedPageRenderer (#541)
