+++ src/Controls/src/Core/HandlerImpl/Application/Application.Tizen.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Maui.Controls
+{
+	public partial class Application
+	{
+	}
+}
