+++ src/Controls/src/Core/HandlerImpl/VisualElement/VisualElement.Tizen.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.Maui.Controls
+{
+	public partial class VisualElement
+	{
+
+	}
+}
